Resolved the issue with process 6 (and probably others) not being added correctly. This bug was due to an error inside of the AddProcessControlBlockToMemoryByProcessID method. Essentially I wasn't passing the correct value for the instructionStart position. I was passing the value for the memory start position when it needed the disk start position.
I also finished the defrag method and tested it. I was going to automatically call the defrag method inside of the method that finds a space for the process; However, I opted out of doing so because I feel this functionality should be controlled by the main controller. (This can be changed back easily if need be)